scsi: Match parameter type to member type.

+++ devices/common/ata/atapibasedevice.h
@@ -45,7 +45,7 @@ public:
     // ScsiPhysInterface methods
     void set_lock_state(bool is_locked) override {};
 
-    void set_xfer_len(uint64_t len) override {
+    void set_xfer_len(int len) override {
         this->xfer_cnt      = len;
         this->r_byte_count  = len;
     }
